DEL: removed leagacy crawler
