+++ src/blueprints/class_blueprint.py
@@ -7,9 +7,7 @@ from flask import Blueprint, request
 from marshmallow import EXCLUDE, ValidationError
 from pymongo.errors import DuplicateKeyError, PyMongoError
 
-from src.common.crawler import get_jupiter_class_infos
 from src.common.database import database
-from src.common.mappers.classes_mapper import break_class_into_events
 from src.common.utils.prettify_preferences import prettify_preferences
 from src.middlewares.auth_middleware import auth_middleware
 from src.repository.user_repository import UserRepository
@@ -110,64 +108,6 @@ def create_class():
     except Exception as ex:
         print(ex)
         return {"message": str(ex)}, 500
-
-
-@class_blueprint.route("many", methods=["POST"])
-@swag_from(f"{yaml_files}/create_many_classes.yml")
-def create_many_classes():
-    try:
-        subject_codes_list = request.json
-        updated = []
-        inserted = []
-        username = request.user.get("Username")
-        logged_user = user_repository.get_by_username(username)
-
-        for subject_code in subject_codes_list:
-            preference_building = None
-            preference_buildings = logged_user.get("buildings")
-            if preference_buildings and len(preference_buildings) > 0:
-                preference_building = str(preference_buildings[0]["_id"])
-
-            subject_classes = get_jupiter_class_infos(subject_code)
-
-            for class_info in subject_classes:
-                class_schema_load = class_schema.load(class_info)
-                events_list = break_class_into_events(
-                    class_schema_load, preference_building
-                )
-
-                for event in events_list:
-                    event_schema_load = event_schema.load(event)
-                    event_schema_load["updated_at"] = datetime.now().strftime(
-                        "%d/%m/%Y %H:%M"
-                    )
-                    event_schema_load["created_by"] = username
-
-                    query = {
-                        "class_code": event_schema_load["class_code"],
-                        "subject_code": event_schema_load["subject_code"],
-                        "week_day": event_schema_load["week_day"],
-                        "created_by": event_schema_load["created_by"],
-                    }
-                    result = events.update_one(
-                        query, {"$set": event_schema_load}, upsert=True
-                    )
-                    updated.append(
-                        event_schema_load["subject_code"]
-                    ) if result.matched_count else inserted.append(
-                        event_schema_load["subject_code"]
-                    )
-
-        return dumps({"updated": updated, "inserted": inserted})
-
-    except Exception as ex:
-        print(ex)
-        return {
-            "message": f"Erro ao buscar informações das turmas - {subject_code}",
-            "updated": updated,
-            "inserted": inserted,
-            "error": str(ex),
-        }, 400
 
 
 @class_blueprint.route("/<subject_code>/<class_code>", methods=["DELETE"])
