--- src/common/new_crawler.py
@@ -1,139 +1,0 @@
-from datetime import datetime
-import requests
-from bs4 import BeautifulSoup, ResultSet
-
-BASE_URL = "https://uspdigital.usp.br/jupiterweb/obterTurma?nomdis=&sgldis="
-CLASS_DIV_IDENTIFIERS = {
-    "style": "border: 2px solid #658CCF; padding: 5px; border-radius: 5px;"
-}
-
-
-class JupiterCrawler:
-    url: str
-    soup: BeautifulSoup
-    classes_divs: ResultSet
-    events: list
-
-    def __init__(self):
-        self.events = []
-
-    def crawl_subject(self, subject_code: str):
-        self.__init__()
-        self.__build_url(subject_code)
-        self.__build_soap()
-        self.__find_classes_divs()
-        self.__extract_classes_info()
-        return self.events
-
-    def __build_url(self, subject_class: str):
-        self.url = BASE_URL + subject_class
-
-    def __build_soap(self):
-        page = requests.get(self.url)
-        self.soup = BeautifulSoup(page.content, "html.parser")
-
-    def __find_classes_divs(self):
-        self.classes_divs = self.soup.find_all("div", attrs=CLASS_DIV_IDENTIFIERS)
-
-    def __extract_classes_info(self):
-        for class_div in self.classes_divs:
-            result = self.__build_events_from_class_div(class_div)
-            self.events += result
-
-    def __build_events_from_class_div(self, class_div):
-        result = []
-        info_tables = class_div.find_all("table")
-        general_info = self.__get_general_info(info_tables)
-        schedule_info_list = self.__get_schedule_info_list(info_tables)
-        student_numbers_info = self.__get_student_numbers_info(info_tables)
-
-        for schedule_info in schedule_info_list:
-            result.append(general_info | schedule_info | student_numbers_info)
-        return result
-
-    def __get_general_info(self, info_tables) -> dict:
-        result = {}
-        general_info_table = info_tables[0]
-        general_info_table_rows = general_info_table.find_all("tr")
-
-        class_code = general_info_table_rows[0].find_all("td")[1].get_text(strip=True)
-        start_period = general_info_table_rows[1].find_all("td")[1].get_text(strip=True)
-        end_period = general_info_table_rows[2].find_all("td")[1].get_text(strip=True)
-        class_type = general_info_table_rows[3].find_all("td")[1].get_text(strip=True)
-        obs = general_info_table_rows[4].find_all("td")[1].get_text(strip=True)
-
-        result["class_code"] = class_code
-        result["start_period"] = start_period
-        result["end_period"] = end_period
-        result["start_period"] = datetime.strptime(start_period, "%d/%m/%Y").strftime(
-            "%Y-%m-%d"
-        )
-        result["end_period"] = datetime.strptime(end_period, "%d/%m/%Y").strftime(
-            "%Y-%m-%d"
-        )
-        result["type"] = class_type
-        result["obs"] = obs
-
-        return result
-
-    def __get_schedule_info_list(self, info_tables) -> list:
-        result = []
-        schedule_info_table = info_tables[1]
-        schedule_info_rows = schedule_info_table.find_all("tr")
-        schedule_info_rows_dropped = schedule_info_rows[1:]
-        schedule_info_enumerate = enumerate(schedule_info_rows_dropped)
-
-        for index, row in schedule_info_enumerate:
-            partial_result = {}
-            data = row.find_all("td")
-
-            week_day = data[0].get_text(strip=True)
-            start_time = data[1].get_text(strip=True)
-            end_time = data[2].get_text(strip=True)
-            teacher = data[3].get_text(strip=True)
-
-            # More than one teacher: only the first row has info, the others only have
-            # the additional teacher
-            if week_day == "" and start_time == "" and end_time == "" and teacher != "":
-                result[index - 1]["teachers"].append(teacher)
-                continue
-
-            partial_result["week_day"] = week_day
-            partial_result["teachers"] = [teacher]
-            partial_result["start_time"] = start_time
-            partial_result["end_time"] = end_time
-
-            result.append(partial_result)
-
-        return result
-
-    def __get_student_numbers_info(self, info_tables) -> dict:
-        result = {
-            "vacancies": 0,
-            "subscribed": 0,
-            "pendings": 0,
-            "enrolled": 0,
-        }
-        student_numbers_table = info_tables[2]
-        student_numbers_rows = student_numbers_table.find_all("tr")
-        student_numbers_rows_dropped = student_numbers_rows[1:]
-
-        filter = {"class": "txt_arial_8pt_black"}
-        for row in student_numbers_rows_dropped:
-            data = row.find_all("span", attrs=filter)
-
-            vacancies_text = data[1].get_text(strip=True)
-            subscribed_text = data[2].get_text(strip=True)
-            pendings_text = data[3].get_text(strip=True)
-            enrolled_text = data[4].get_text(strip=True)
-
-            if vacancies_text != "" and vacancies_text.isdigit():
-                result["vacancies"] += int(vacancies_text)
-            if subscribed_text != "" and subscribed_text.isdigit():
-                result["subscribed"] += int(subscribed_text)
-            if pendings_text != "" and pendings_text.isdigit():
-                result["pendings"] += int(pendings_text)
-            if enrolled_text != "" and enrolled_text.isdigit():
-                result["enrolled"] += int(enrolled_text)
-
-        return result
